backend/requirements.dev.txt
Django==3.2.9
djangorestframework==3.11.0
django-cors-headers==3.13.0
pytz==2022.1
djoser==2.1.0
python-dotenv==0.20.0
pillow==9.2.0
django-filter==22.1

celery==5.2.7
django-celery-results==2.4.0
redis==4.3.4
django-redis==5.2.0
django-celery-beat==2.3.0
flower==1.2.0

psycopg2-binary==2.9.3